+++ tfhe/src/shortint/server_key/scalar_sub.rs
@@ -117,7 +117,7 @@ impl ServerKey {
     /// assert_eq!(msg - scalar as u64, clear);
     /// ```
     pub fn scalar_sub_assign(&self, ct: &mut Ciphertext, scalar: u8) {
-        let modulus = self.message_modulus.0 as u64;
+        let modulus = self.message_modulus.0;
         let acc = self.generate_lookup_table(|x| (x.wrapping_sub(scalar as u64)) % modulus);
         self.apply_lookup_table_assign(ct, &acc);
     }
@@ -207,13 +207,13 @@ impl ServerKey {
     pub fn unchecked_scalar_sub_assign(&self, ct: &mut Ciphertext, scalar: u8) {
         let neg_scalar = neg_scalar(scalar, ct.message_modulus);
 
-        let delta = (1_u64 << 63) / (self.message_modulus.0 * self.carry_modulus.0) as u64;
-        let shift_plaintext = neg_scalar as u64 * delta;
+        let delta = (1_u64 << 63) / (self.message_modulus.0 * self.carry_modulus.0);
+        let shift_plaintext = u64::from(neg_scalar) * delta;
         let encoded_scalar = Plaintext(shift_plaintext);
 
         lwe_ciphertext_plaintext_add_assign(&mut ct.ct, encoded_scalar);
 
-        ct.degree += Degree::new(neg_scalar as usize);
+        ct.degree += Degree::new(u64::from(neg_scalar));
     }
 
     pub fn unchecked_scalar_sub_assign_with_correcting_term(
@@ -222,19 +222,19 @@ impl ServerKey {
         scalar: u8,
     ) {
         let msg_mod = self.message_modulus.0;
-        assert!((scalar as usize) < msg_mod);
+        assert!((u64::from(scalar)) < msg_mod);
 
-        let delta = (1_u64 << 63) / (self.message_modulus.0 * self.carry_modulus.0) as u64;
+        let delta = (1_u64 << 63) / (self.message_modulus.0 * self.carry_modulus.0);
 
         let encoded_scalar = Plaintext(scalar as u64 * delta);
         lwe_ciphertext_plaintext_sub_assign(&mut ct.ct, encoded_scalar);
 
         let correcting_term = ct.degree.get().div_ceil(msg_mod).max(1) * msg_mod;
-        let encoded_msg_mod = Plaintext(correcting_term as u64 * delta);
+        let encoded_msg_mod = Plaintext(correcting_term * delta);
         lwe_ciphertext_plaintext_add_assign(&mut ct.ct, encoded_msg_mod);
 
         // subtracted scalar, added the correcting term.
-        ct.degree += Degree::new(correcting_term - scalar as usize);
+        ct.degree += Degree::new(correcting_term - u64::from(scalar));
         // noise does not change as operations only involved plaintexts
     }
 
@@ -475,7 +475,7 @@ impl ServerKey {
 }
 
 fn neg_scalar(scalar: u8, msg_modulus: MessageModulus) -> u8 {
-    let msg_modulus = msg_modulus.0 as u64;
+    let msg_modulus = msg_modulus.0;
 
     let scalar = scalar as u64 % msg_modulus;
